requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
cycler==0.10.0
Django==3.2.7
gunicorn==20.1.0
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.3
Naked==0.1.31
numpy==1.21.2
Pillow==8.3.2
pycryptodome==3.11.0
pyparsing==2.4.7
pypng==0.0.21
PyQRCode==1.2.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
qrcode==7.3
requests==2.26.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
